Enhance error handling
We stumbled on DNS denial dued to adblockers, here is the fix for the lib to stay consistent. I you could do a haxelib release it would be super awesome.

We also added a flag to silence the lib on purpose (you can refuse this one but it is practical.

Thanks you for your awesome work.